+++ lib/src/rules_factory.dart
@@ -6,6 +6,7 @@ import 'rules/newline_before_return.dart';
 import 'rules/no_boolean_literal_compare_rule.dart';
 import 'rules/no_empty_block.dart';
 import 'rules/no_magic_number_rule.dart';
+import 'rules/no_object_declaration.dart';
 import 'rules/prefer_intl_name.dart';
 import 'rules/prefer_trailing_comma_for_collection.dart';
 
@@ -17,6 +18,7 @@ const _implementedRules = {
   NoBooleanLiteralCompareRule(),
   NoEmptyBlockRule(),
   NoMagicNumberRule(),
+  NoObjectDeclarationRule(),
   PreferIntlNameRule(),
   PreferTrailingCommaForCollectionRule(),
 };