+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
+# Unreleased
+
+- Add static code diagnostic prefer-conditional-expressions
+
 # 1.7.1
+
 - Support analyzer_plugin 0.3.0
 
 # 1.7.0
+
 - Add experimental static code diagnostics binary-expression-operand-order, prefer-intl-name
 - Add Number of Methods metric
 - Drop dependency on resource
@@ -11,56 +17,71 @@
 - Set min SDK version to >=2.6.0.
 
 # 1.6.0
+
 - Add experimental static code diagnostics prefer-trailing-comma-for-collection, no-magic-number
 - Support number of arguments metric in analyzer plugin
 - Support excluding files from metrics calculation
 
 # 1.5.1
+
 - Improve code diagnostics double-literal-format, no-boolean-literal-compare
 - Add experimental static code diagnostics newline-before-return, no-empty-block, avoid-preserve-whitespace-false
 - Support cyclomatic complexity metric in analyzer plugin
 
 # 1.5.0
+
 - Add experimental static code diagnostics double-literal-format, no-boolean-literal-compare
 
 # 1.4.0
+
 - Drop dependency on built_collection
 - Add set-exit-on-violation-level cli argument
 
 # 1.3.1
+
 - Fix get arguments count
 
 # 1.3.0
+
 - Add Number of Arguments metrics
 
 # 1.2.1
+
 - Validate root-folder argument
 - Fix paths to analyze fail to validate with non-default root-folder
 - Fix paths weren't validated to be inside root-folder
 - Support factory constructors analysis
 
 # 1.2.0
+
 - Allow to analyze multiple directories
 
 ## 1.1.5
+
 - Tweak console reporter
 
 ## 1.1.4
+
 - Add some dartdocs
 - Update readme
 - Add library usage example
 
 ## 1.1.3
+
 - Fix validate input arguments
 
 ## 1.1.2
+
 - Improve CodeClimate report
 
 ## 1.1.1
+
 - Added support extension methods
 
 ## 1.1.0
+
 - Added support for CodeClimate
 
 ## 1.0.0
+
 - Initial release