+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # Unreleased
 
 - Add static code diagnostic prefer-conditional-expressions
-- Add static code diagnostic  no-object-declaration
+- Add static code diagnostic no-object-declaration
 
 # 1.7.1